fix(ENG-980): desktop version of forget password
