+++ src/pages/forget-password/otp/desktop/desktop.module.scss
@@ -1,0 +1,53 @@
+.container {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--color-off-white-01);
+
+  .card {
+    width: 33.75rem;
+    padding: 1rem 0;
+  }
+
+  .header {
+    padding: 1rem 1.5rem;
+  }
+
+  .icon {
+    cursor: pointer;
+  }
+
+  .main {
+    display: flex;
+    flex-direction: column;
+    gap: 2.5rem;
+    padding: 4.5rem;
+  }
+
+  .content {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+  }
+
+  .title {
+    font-size: 1.5rem;
+    font-weight: 600;
+  }
+
+  .text {
+    font-size: 1rem;
+    font-weight: 400;
+  }
+
+  .footer {
+    padding: 1rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0.5rem;
+    border-top: 1px solid var(--border-gray-01);
+  }
+}
