+++ src/pages/forget-password/otp/mobile/mobile.module.scss
@@ -1,0 +1,45 @@
+.container {
+  background-color: #ffffff;
+  height: 100vh;
+  display: grid;
+  grid-template-rows: 6rem 1fr 9.75rem;
+}
+
+.header {
+  padding: 5rem 1.43rem 1.43rem 1.43rem;
+}
+
+.main {
+  margin-top: 2.5rem;
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.title {
+  font-size: 1.5rem;
+  font-weight: 600;
+}
+
+.text {
+  font-size: 1rem;
+  font-weight: 400;
+}
+
+.footer {
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  border-top: 1px solid var(--border-gray-01);
+}
+
+.otp {
+  margin-top: 2.5rem;
+}
