+++ src/pages/forget-password/password/desktop/desktop.module.scss
@@ -1,0 +1,46 @@
+.container {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--color-off-white-01);
+
+  .card {
+    width: 33.75rem;
+    padding: 1rem 0;
+  }
+
+  .header {
+    padding: 1rem 1.5rem;
+  }
+
+  .icon {
+    cursor: pointer;
+  }
+
+  .main {
+    display: flex;
+    flex-direction: column;
+    gap: 2.5rem;
+    padding: 4.5rem;
+  }
+
+  .title {
+    font-size: 1.5rem;
+    font-weight: 600;
+  }
+
+  .footer {
+    padding: 1rem;
+    display: flex;
+    justify-content: center;
+    border-top: 1px solid var(--border-gray-01);
+  }
+
+  .newPassword {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+}
