+++ src/pages/forget-password/password/mobile/mobile.module.scss
@@ -2,7 +2,7 @@
   background-color: #ffffff;
   height: 100vh;
   display: grid;
-  grid-template-rows: 6rem 1fr 9.75rem;
+  grid-template-rows: 6rem 1fr 6.75rem;
 }
 
 .header {
@@ -16,30 +16,19 @@
   padding: 2rem;
 }
 
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
 .title {
   font-size: 1.5rem;
-  font-weight: 600px;
-}
-
-.text {
-  font-size: 1rem;
-  font-weight: 400;
+  font-weight: 600;
+  margin-bottom: 2.5rem;
 }
 
 .footer {
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
+  padding: 1rem 1rem 0 1rem;
   border-top: 1px solid var(--border-gray-01);
 }
 
-.otp {
-  margin-top: 2.5rem;
+.newPassword {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
 }
